ファイル名とクラス名が不適切だったので修正
以下のように修正した
memo.js(MemoAppクラス) -> MemoApp.js(MemoAppクラス)
dbMemo.js(DbMemoクラス) -> Memo.js(Memoクラス)
